--- card-moved.js
@@ -1,25 +1,0 @@
-const { addLabel, getIssueFromCard, moveZubeCard } = require('./shared');
-const { addLoggingToRequest } = require('./logger');
-const { client, lock } = require('./redis');
-
-module.exports = async function onCardMoved(context) {
-  addLoggingToRequest(context);
-  const result = await getIssueFromCard(context);
-  if (result) {
-    const {
-      issue,
-      column: { name },
-    } = result;
-    // move to different board in Zube
-    const label = `[zube]: ${name}`;
-    // const redis = client();
-    // await lock(redis, issue.number);
-    await addLabel({
-      context,
-      issue,
-      existingLabelRegex: /\[zube\]:/,
-      newLabel: label,
-    });
-    await moveZubeCard(context, result);
-  }
-};
