+++ test/fixtures/github/issue-diff-project.json
@@ -10,7 +10,7 @@
                 "name": "In progress"
               },
               "project": {
-                "name": "Team Rocket",
+                "name": "Team Board",
                 "columns": {
                   "nodes": [
                     {
