+++ test/pre-push.sh
@@ -1,0 +1,1 @@
+yarn test
